LayoutTests/fast/forms/validation-message-user-modify-expected.txt

The inner element of validation should not be modifiable.


PASS renderTreeAfterEditing is renderTreeBeforeEditing
PASS successfullyParsed is true

TEST COMPLETE
